requirements.txt
absl-py==2.1.0
altair==5.3.0
astunparse==1.6.3
attrs==23.2.0
blinker==1.8.2
Bottleneck==1.3.7
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.11.0
dill==0.3.8
filelock==3.14.0
flatbuffers==24.3.25
fonttools==4.51.0
fsspec==2024.5.0
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.64.0
h5py==3.11.0
idna==3.7
importlib_metadata==7.1.0
importlib-resources==6.1.1
intel-openmp==2021.4.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
keras==3.3.3
kiwisolver==1.4.4
lapx==0.5.9
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
mkl==2021.4.0
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
ml-dtypes==0.3.2
mpmath==1.3.0
namex==0.0.8
networkx==3.2.1
numexpr==2.8.7
numpy==1.26.4
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
packaging==23.2
pandas==2.2.1
pillow==10.3.0
pip==24.0
ply==3.11
protobuf==4.25.3
psutil==5.9.8
py-cpuinfo==9.0.0
pyarrow==16.1.0
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.0.9
PyQt5==5.15.10
PyQt5-sip==12.13.0
python-dateutil==2.8.2
pytube==15.0.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.2
rich==13.7.1
rpds-py==0.18.1
scipy==1.13.0
seaborn==0.13.2
setuptools==69.5.1
sip==6.7.12
six==1.16.0
smmap==5.0.1
streamlit==1.34.0
sympy==1.12
tbb==2021.12.0
tenacity==8.3.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
thop==0.1.1.post2209072238
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.3.0
torchvision==0.18.0
tornado==6.3.3
tqdm==4.66.4
typing_extensions==4.11.0
tzdata==2023.3
ultralytics==8.2.19
unicodedata2==15.1.0
urllib3==2.2.1
watchdog==4.0.0
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
zipp==3.17.0
